Disable gem installers
There is no need to run this script any longer. Its better to just use Homebrew and rbenv. Manually install the gems you need at the time you need them for the env you need them in.

I'm keeping the gems commented out for reference on which gems are useful to me.